Extract language code into separate module
