+++ src/language.h
@@ -1,0 +1,12 @@
+#pragma once
+#ifndef CATA_SRC_LANGUAGE_H
+#define CATA_SRC_LANGUAGE_H
+
+#include <string>
+
+bool isValidLanguage( const std::string &lang );
+std::string getLangFromLCID( const int &lcid );
+void select_language();
+void set_language();
+
+#endif // CATA_SRC_LANGUAGE_H
